# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Safad Vibe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search/history:
    get:
      operationId: getSearchHistory
      tags: [search]
      summary: Get recent search history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [products, flights, hotels, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchHistoryResponse"
...
components:
  schemas:
...
    SearchHistoryItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [products, flights, hotels]
        query:
          type: string
        createdAt:
          type: string
          format: date-time
        resultCount:
          type: integer
      required:
        - id
        - type
        - query
        - createdAt
    SearchHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchHistoryItem"
        total:
          type: integer
      required:
        - items
        - total
...