# lib/api-zod/src/generated/types/searchHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Safad Vibe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchHistoryItemType } from "./searchHistoryItemType";

export interface SearchHistoryItem {
  id: string;
  type: SearchHistoryItemType;
  query: string;
  createdAt: Date;
  resultCount?: number;
}